requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
dnspython
kafka-python
requests
